+++ R/add_aks_methods.R
@@ -22,7 +22,7 @@
 #' - `enable_rbac`: Whether to enable role-based access controls.
 #' - `agent_pools`: A list of pool specifications. See 'Details'.
 #' - `login_user,login_passkey`: Optionally, a login username and public key (on Linux). Specify these if you want to be able to ssh into the cluster nodes.
-#' - `cluster_service_principal`: The service principal (client) that AKS will use to manage the cluster resources. This should be a list, with the first component being the client ID and the second the client secret. If not supplied, the values are obtained from the service principal used for this ARM login.
+#' - `cluster_service_principal`: The service principal (client) that AKS will use to manage the cluster resources. This should be a list, with the first component being the client ID and the second the client secret. If not supplied, a new service principal will be created (requires an interactive session).
 #' - `properties`: A named list of further Kubernetes-specific properties to pass to the initialization function.
 #' - `wait`: Whether to wait until the AKS resource provisioning is complete. Note that provisioning a Kubernetes cluster can take several minutes.
 #' - `...`: Other named arguments to pass to the initialization function.
@@ -229,7 +229,7 @@ add_aks_methods <- function()
             )
 
         if(is.null(cluster_service_principal))
-            cluster_service_principal <- get_app_details(self$token)
+            cluster_service_principal <- create_service_principal(name, location)
 
         if(is.null(props$servicePrincipalProfile))
             props$servicePrincipalProfile <- list(
@@ -327,3 +327,14 @@ add_aks_methods <- function()
     })
 }
 
+
+create_service_principal <- function(name, location)
+{
+    gr <- try(AzureGraph::get_azure_login(), silent=TRUE)
+    if(inherits(gr, "try-error"))
+        gr <- AzureGraph::create_azure_login()
+    
+    appname <- paste("RAKSapp", name, location, sep="-") 
+    app <- gr$create_app(appname)
+    list(app$properties$appId, app$password)
+}